fix(ui): heartbeat detects going offline too, not only recovery
